facecify-master/requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
itsdangerous==2.2.0
joblib==1.4.2
keras-facenet==0.3.2
mtcnn==0.1.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-contrib-python-headless==4.10.0.84
opencv-python==4.10.0.84
python-dotenv==1.0.1
scikit-learn==1.5.1
tensorflow==2.17.0
Werkzeug==3.0.4
WTForms==3.1.2
psycopg2==2.9.9
email-validator==2.2.0
gunicorn==21.2.0
